feat(airc-lib): expose work coordination API
