+++ crates/airc-lib/src/room.rs
@@ -185,76 +185,9 @@ fn set_owner_only_permissions(_path: &Path) -> std::io::Result<()> {
 fn now_ms() -> u64 {
     std::time::SystemTime::now()
         .duration_since(std::time::UNIX_EPOCH)
-        .map(|d| d.as_millis() as u64)
-        .unwrap_or(0)
+        .expect("system clock before UNIX_EPOCH violates AIRC room timestamp contract")
+        .as_millis() as u64
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use tempfile::TempDir;
-
-    #[test]
-    fn from_name_is_deterministic_across_homes() {
-        // Same room name → same channel UUID across different homes.
-        // Critical for peers joining the same name without sharing
-        // the channel UUID out-of-band.
-        let home_a = TempDir::new().unwrap();
-        let home_b = TempDir::new().unwrap();
-        let a = Room::from_name(home_a.path(), "project-x");
-        let b = Room::from_name(home_b.path(), "project-x");
-        assert_eq!(a.channel, b.channel);
-        // Wire IS different (per-home), but channel matches.
-        assert_ne!(a.wire, b.wire);
-    }
-
-    #[test]
-    fn from_name_differs_per_name() {
-        let home = TempDir::new().unwrap();
-        let a = Room::from_name(home.path(), "general");
-        let b = Room::from_name(home.path(), "private");
-        assert_ne!(a.channel, b.channel);
-        assert_ne!(a.wire, b.wire);
-    }
-
-    #[test]
-    fn load_or_default_returns_default_when_missing() {
-        let home = TempDir::new().unwrap();
-        let room = load_or_default(home.path()).unwrap();
-        assert_eq!(room.name, "default");
-    }
-
-    #[test]
-    fn save_then_load_roundtrips() {
-        let home = TempDir::new().unwrap();
-        let original = Room::from_name(home.path(), "test-room");
-        save(home.path(), &original).unwrap();
-        let loaded = load_or_default(home.path()).unwrap();
-        assert_eq!(loaded.name, original.name);
-        assert_eq!(loaded.channel, original.channel);
-        assert_eq!(loaded.wire, original.wire);
-    }
-
-    #[test]
-    fn sanitise_replaces_path_separators() {
-        // Avoid escaping the wires dir via weird names.
-        assert_eq!(sanitise_name("../etc/passwd"), "---etc-passwd");
-        assert_eq!(sanitise_name("normal-name_42"), "normal-name_42");
-    }
-
-    #[test]
-    fn refuses_unknown_schema_version() {
-        let home = TempDir::new().unwrap();
-        std::fs::create_dir_all(home.path()).unwrap();
-        std::fs::write(
-            path_in(home.path()),
-            r#"{"version":999,"name":"x","wire":"/tmp","channel":"00000000-0000-0000-0000-000000000000","joined_at_ms":0}"#,
-        )
-        .unwrap();
-        let result = load_or_default(home.path());
-        assert!(matches!(
-            result,
-            Err(RoomError::SchemaVersionMismatch { found: 999, .. })
-        ));
-    }
-}
+mod tests;
