+++ crates/airc-lib/src/time.rs
@@ -1,0 +1,6 @@
+pub(crate) fn now_ms() -> u64 {
+    std::time::SystemTime::now()
+        .duration_since(std::time::UNIX_EPOCH)
+        .expect("system clock before UNIX_EPOCH violates AIRC timestamp contract")
+        .as_millis() as u64
+}
